feat: 리스트업 수정 기능 개발

+++ src/main/java/be15fintomatokatchupbe/campaign/command/domain/aggregate/entity/PipelineStep.java
@@ -13,9 +13,9 @@ public class PipelineStep {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer pipelineStepId;
+    private Long pipelineStepId;
 
-    private Integer level;
+    private Long level;
 
     private String stepName;
 }
